+++ pyogrio/geopandas.py
@@ -156,12 +156,13 @@ def read_dataframe(
         ``max_features``, ``where``, ``bbox``, ``mask``, or ``sql``). Note that
         the starting index is driver and file specific (e.g. typically 0 for
         Shapefile and 1 for GeoPackage, but can still depend on the specific
-        file). The performance of reading (a large number of) features usings FIDs
-        is also driver specific and dependent on the value of ``use_arrow``. The order
+        file). The performance of reading a large number of features usings FIDs
+        is also driver specific and depends on the value of ``use_arrow``. The order
         of the rows returned is undefined. If you would like to sort based on FID, use
         ``fid_as_index=True`` to have the index of the GeoDataFrame returned set to the
         FIDs of the features read. If ``use_arrow=True``, the number of FIDs is limited
-        to 4997 for drivers with 'OGRSQL' as default SQL dialect.
+        to 4997 for drivers with 'OGRSQL' as default SQL dialect. To read a larger
+        number of FIDs, set ``user_arrow=False``.
     sql : str, optional (default: None)
         The SQL statement to execute. Look at the sql_dialect parameter for more
         information on the syntax to use for the query. When combined with other