+++ pyogrio/tests/test_geopandas_io.py
@@ -514,7 +514,7 @@ def test_read_fids(naturalearth_lowres_all_ext, fids, use_arrow):
 
 
 @requires_arrow_api
-def test_read_fids_arrow_many_error(naturalearth_lowres):
+def test_read_fids_arrow_max_exception(naturalearth_lowres):
     # Maximum number at time of writing is 4997 for "OGRSQL". For e.g. for SQLite based
     # formats like Geopackage, there is no limit.
     nb_fids = 4998